+++ src/RDD.Web.AutoMapper/RddBuilderExtensions.cs
@@ -1,0 +1,18 @@
+﻿using AutoMapper;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Rdd.Web.Helpers;
+using Rdd.Web.Querying;
+
+namespace Rdd.Web.AutoMapper
+{
+    public static class RddBuilderExtensions
+    {
+        public static RddBuilder AddAutoMapper(this RddBuilder rddBuilder)
+        {
+            rddBuilder.Services.TryAddSingleton(typeof(IRddObjectsMapper<,>), typeof(RddObjectsMapper<,>));
+            rddBuilder.Services.TryAddSingleton(Mapper.Instance);
+
+            return rddBuilder;
+        }
+    }
+}