+++ test/Rdd.Web.Tests/Models/Cat.cs
@@ -1,0 +1,18 @@
+﻿using Rdd.Domain;
+
+namespace Rdd.Web.Tests.Models
+{
+    public class Cat : IEntityBase<int>
+    {
+        public string Name { get; set; }
+
+        public string Url { get; }
+
+        public int Id { get; set; }
+
+        public int Age { get; set; }
+
+        public Cat() { }
+        public object GetId() => Id;
+    }
+}